[finishes #171810827] refactoring
