+++ tests/integration/adam_test.go
@@ -8,7 +8,6 @@ import (
 	"github.com/itmo-eve/eden/pkg/einfo"
 	"github.com/itmo-eve/eden/pkg/elog"
 	"github.com/itmo-eve/eden/pkg/utils"
-	"github.com/lf-edge/eve/api/go/info"
 	"os"
 	"path"
 	"path/filepath"
@@ -107,21 +106,8 @@ func TestAdamLogs(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	done := make(chan error)
-	go func() {
-		done <- elog.LogWatchWithTimeout(ctx.GetLogsDir(devUUID), map[string]string{"devId": devUUID.String()}, elog.HandleFirst, 600)
-	}()
-	go func() {
-		handler := func(item *elog.LogItem) bool {
-			done <- nil
-			return elog.HandleFirst(item)
-		}
-		err := elog.LogLast(ctx.GetLogsDir(devUUID), map[string]string{"devId": devUUID.String()}, handler)
-		if err != nil {
-			done <- err
-		}
-	}()
-	if <-done != nil {
+	err = elog.LogChecker(ctx.GetLogsDir(devUUID), map[string]string{"devId": devUUID.String()}, 600)
+	if err != nil {
 		t.Fatal(err)
 	}
 }
@@ -131,24 +117,7 @@ func TestAdamInfo(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	done := make(chan error)
-
-	go func() {
-		err = einfo.InfoWatchWithTimeout(ctx.GetInfoDir(devUUID), map[string]string{"devId": devUUID.String()}, einfo.ZInfoDevSWFind, einfo.HandleFirst, 300)
-		done <- err
-	}()
-	go func() {
-		handler := func(im *info.ZInfoMsg, ds []*info.ZInfoDevSW) bool {
-			einfo.ZInfoDevSWPrn(im, ds)
-			done <- nil
-			return true
-		}
-		err = einfo.InfoLast(ctx.GetInfoDir(devUUID), map[string]string{"devId": devUUID.String()}, einfo.ZInfoDevSWFind, handler)
-		if err != nil {
-			done <- err
-		}
-	}()
-	err = <-done
+	err = einfo.InfoChecker(ctx.GetInfoDir(devUUID), map[string]string{"devId": devUUID.String()}, 300)
 	if err != nil {
 		t.Fatal(err)
 	}
